Add --end flag for add command
Instead of specifying an event duration during creation, it can be
useful to give an event end time. This commit adds a --end flag that can
be used as an alternative to the --duration flag.

The --end flag takes a date string similar to the --when flag.

Given both a --end flag and a --duration flag, the duration value
will take precedence. Given neither flag, duration will be asked for
in interactive mode.