Testing framework setup [fixes #343]
In order to get testing working nicely with pytest and tox, I had to go
ahead and add a package directory going and make some changes to
setup.py

Setup flake8 to run with tox and therefore linted everything.

I'd like to drop support for 2.6, I've removed it from the setup.py file.