Added man1 page to docs/man1
I used the text of the very long __doc__ string to create a man page.
The main usage message is now generated via argparse, so I thought it
was time to move the __doc__ string to some other location.  Since
debian maintainers are asking for a proper manpage, I figured this would
be a helpful thing to do.