Rely on arparse to provide defaults
Since argparse can provide defaults and destinations, we don't need to
assign to special names within the GoogleCalendarInterface class.  We
can just **kwargs and use that dictionary to store the values.

In testing, for initializing the class, we can use the subparsers
against the empty strings to pass in any defaults that the code needs.

Definitely need more tests.

Also, I probably broke the passing along to GoogleAuth because (I
suppose) it expects a namespace rather than a dict.  We need to test
that.  I'll make an issue.