Default to prompting [fixes #348]
We lost the --[no]flag idiom with the move to argparse.  We really ought
to go over the options with a fine tooth comb.  One that stood out to me
was not prompting by default.  This readds default prompting on the
'add' command and allows supression of prompting with --noprompt.