Cast opts back to Namespace [Fixes #356]
When we moved to using a dictionary of options in the gcal instance we
suddently couldn't directly pass a dictionary to run_flow --- it expects
an argparse namespace.  For now, we just recreate the namespace when we
call run_flow.  I say 'for now', because I need to closely examine the
authentication process, the library versions, etc---we have a ticket for
that work.

I added a test to test_gcalcli, but it's  not actually relevant to this
fix.  I'm leaving it in this commit because it doesn't hurt.