Refactor utils: remove unneeded class structure
We can import the parsedatetime module in the utils namespace and use it
directly in the function which was called fromString.   We will only
ever need a single instance of this parser per run (as far as I can
tell) so there's no need to instantiate a DateTimeParser both when we do
get_time_from_str and when we initialize the GoogleCalendarInteface
object.

Secondly, the erstwhile get_time_from_str was really about getting a
start and end time given a duration.  So, I renamed it it to
`get_times_from_duration`.  And then the erstwhile
DateTimeParser.fromString is the new utils.get_time_from_str