Require parseddatetime, fix several unreported bugs
Explicitly require parsedatetime for fuzzy matching of dates.  This is
available in pip and can therefore be added to the setup.py.  We can
roll this back if there's some reason that we can't require this
package.

Added tests to demonstrate functionality.  Now explicitly try
dateutil.parser.parse and then fall back on parseddatetime.  I cleaned
up the naming in utils to try to make things clearer.

Similarly, added an actual raise when we needed one and removed some
try/excepts which were allowing us to fail without raising.  Our tests
were passing because we emitted an error message and then returned
instead of raising the error.

I found a handful of bugs which I introducted when trying to refactor
the argument parser stuff.  Too much copypasta!