Test improvements
We need to return meaningful values from our functions and assert things
about our tests.

In general, I'm moving towards an idea that GoogleCalendarInterface
should raise errors rather than printing messages and returning None.
Then, the CLI code (in main) should except these errors, and print the
messages and exit with error code.  The advantage of this is twofold,
GoogleCalendarInterface become programmatically useful for import and,
relatedly, we can write tests that assert return values against the
methods of GoogleCalendarInterface.

This is a step in the right direction, I think, but still a work in
progress.