parsedatetime is not an optional requirement
Since v4.0.0 parsedatetime has been a requirement. [1] parsedatetime was
originally an "extra", a recommended optional package. The commit adding
it as a hard dependency didn't remove it from the list of extras.

Before this change both docs/README.md and setup.py list parsedatetime
twice. After this change parsedatetime is not described as optional.

[1] https://github.com/insanum/gcalcli/commit/166a58a81068cef858caece10e41db591f98ac53