Make all_cals count errors more explicit
When something goes wrong with parsing your --calendar option (or your
defaults are wrong, or whatever) and you end up with 0 calendars
available, in the context where the calendar count has to be == 1, we
say "You must select only one calendar".  But this is confusing because
the problem is not that you have selected > 1, but that you ended up
with 0.  Hopefully this messaging change helps people realize what went
wrong faster.