refactor: move _tsv() implementation to gcalcli.details
Refactor `gcalcli.gcal.GoogleCalendarInterface._tsv()`, replacing the
long list of if statements with a modular design using a series of
`gcalcli.details.Handler` classes.

Each `Handler` represents the set of event properties that are
captured by each item of `gcalcli.argparsers.DETAILS`. So far only the
`header` class attribute and the `get()` methods are implemented. This
is what is necessary for TSV output.

Preparatory work for #550.