fix: change SimpleSingleColumnHandler.header to a simple list
Can't use a property on a classmethod without making a custom
metaclass. Without that, the property is only enacted upon class
instantiation, which we aren't doing here.

A custom metaclass is too magic, and thus far we have avoided
instantiating what would be singleton objects.

Also changed `SimpleSingleColumnHandler._get()` to use `cls.header[0]`
instead of `cls._key` which is removed.