feat: add agendaupdate for time
Round trips don't work for all day events. The current output format
is to present them as events starting and ending at midnight of the
same day. But this keeps us from having events starting and ending at
midnight of the same day.

In the next commit I plan to change the output for all day events so
the time fields are blank.

Also:

- change signature of `Handler.patch()` to include `cal`, which is necessary to extract the timezone
- change default of `Handler.fieldnames` from `None` to `[]` to silence a mypy warning.

Partially implements #550.