feat: print an error message when ~/.gcalcli_oauth deserialization fails
At the moment, only a backtrace is printed and it doesn't make what's
going on obvious.

This is especially important now as the migration from oauth2client to
google-auth comes with a format change for this file and anyone who had
been using gcalcli successfully before is likely to encounter this error.